Add assert_shape() routine that asserts a given input matrix shape. Use the routine in other linalg routines.
Some matrix operations are only defined for specific shapes of the
matrices they are applied to. E.g., the determinant makes sense only
for square matrices. In other cases, they shape of input and output
matrices can be complex, as in the case of the singular value
decomposition. In this commit, tests for the shape of input and
possibly output matrices are introduced for the routines det(), inv(),
solve(), svd(), eigvals(), eig() and eigh(). The tests give
runtime errors and may stop the program, which is better than running
the program with wrong or meaningless data. Hopefully this approach
helps avoiding possible problems with assumed shape arrays.